Update Get-InputApiUrl.ps1
Update the working logic of the Get-InputApiUrl to support LogRhythm Cloud operating over port 443 in place of the pre-configured 8501.  The Cmdlet is more universal in retaining the port if the port is existing in the configuration file but also can be updated to a new user-provided port if it conforms to hostname:port convention.