Change name search behavior
The Alarms API name property does not behave like other RestAPI cmdlet behavior (Lists, Cases, Users) in that it does not perform a wildcard lookup.

The result is the cmdlet performs the name matching after all results are collected (after applying any additional filter criteria).  If the -exact flag is provided then the Name attribute is submitted to the RestAPI.