Update Get-LrCases.ps1
Adds new capability for retrieving cases based on tags, whereby you can select to return cases with all tags requested, or any of the tags requested.

Note with any tags, if no tags out of the tags listed as assigned to the case(s) then no results will be returned.  At least one of the requested tags is required to be present.