Update Response check if Error present
Discovered scenario where Response contained legitimate data and without any type of pagination still would trigger the error block to catch.  Improved logic to check if Response.Error is not null and that Response.Error is true.

This resolved the presence of the anomalous data return.  Root cause is still unknown, will log an issue if it re-presents itself.

Originally identified on Get-LrIdentities.